requirements.txt
transformers==4.49.0
torch==2.6.0
torchvision==0.21.0
opencv-python==4.11.0.86
pillow==11.1.0
pytest==8.3.4
